Moving metrics tests
